move controllers folder into src
